+++ nengo_spa/__init__.py
@@ -22,6 +22,7 @@ from nengo_spa.operators import dot, reinterpret, translate
 from nengo_spa.semantic_pointer import SemanticPointer
 from nengo_spa.vector_generation import (
     AxisAlignedVectors,
+    EquallySpacedPositiveUnitaryHrrVectors,
     ExpectedUnitLengthVectors,
     OrthonormalVectors,
     UnitaryVectors,